# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/replies/text": {
      "post": {
        "operationId": "RepliesController_generateText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReplyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI-generated text response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/replies/speech": {
      "post": {
        "operationId": "RepliesController_generateSpeech",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextToSpeechDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Speech audio in Base64",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextToSpeechResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/replies/speech/timestamped": {
      "post": {
        "operationId": "RepliesController_generateTimestampedSpeech",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextToSpeechTimestampedDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Timestamped audio with Base64 encoding",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextToSpeechTimestampedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/replies/full/plain": {
      "post": {
        "operationId": "RepliesController_generateFullPlain",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReplyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text and TTS audio",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullReplyPlainResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/replies/full/timestamped": {
      "post": {
        "operationId": "RepliesController_generateFullTimestamped",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReplyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text and timestamped speech",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FullReplyTimestampedResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/replies/speech/realtime": {
      "post": {
        "operationId": "RepliesController_realtimeVoice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RealtimeVoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "WebRTC answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebRtcAnswerResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/replies/transcription/realtime": {
      "post": {
        "operationId": "RepliesController_realtimeTranscription",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RealtimeTranscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transcription session details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionSessionCreateResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/replies/providers": {
      "get": {
        "operationId": "RepliesController_getProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of providers with availability status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AiProviderStatusDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "replies"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Classroom Conversation Partner API",
    "description": "NestJS powered API with OpenAPI documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GenerateReplyDto": {
        "type": "object",
        "properties": {
          "inputText": {
            "type": "string",
            "description": "User input text"
          },
          "previousMessages": {
            "description": "Previous conversation messages",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "personality": {
            "type": "object",
            "description": "Personality configuration"
          },
          "conversationRole": {
            "type": "object",
            "description": "Conversation role"
          },
          "language": {
            "type": "string",
            "description": "Language setting"
          },
          "scenario": {
            "type": "object",
            "description": "Scenario configuration"
          },
          "userProfile": {
            "type": "object",
            "description": "User profile information"
          }
        },
        "required": [
          "inputText"
        ]
      },
      "TextToSpeechDto": {
        "type": "object",
        "properties": {
          "inputMessage": {
            "type": "string",
            "description": "Text to convert to speech"
          },
          "personality": {
            "type": "object",
            "description": "Personality for voice settings"
          },
          "language": {
            "type": "string",
            "description": "Language setting"
          },
          "responseFormat": {
            "type": "string",
            "description": "Audio response format"
          }
        },
        "required": [
          "inputMessage"
        ]
      },
      "TextToSpeechResponseDto": {
        "type": "object",
        "properties": {
          "audioBase64": {
            "type": "string",
            "description": "Base64 encoded audio"
          },
          "sampleRate": {
            "type": "number",
            "description": "Sample rate of the audio"
          }
        },
        "required": [
          "audioBase64",
          "sampleRate"
        ]
      },
      "TextToSpeechTimestampedDto": {
        "type": "object",
        "properties": {
          "inputMessage": {
            "type": "string",
            "description": "Text to convert to speech"
          },
          "personality": {
            "type": "object",
            "description": "Personality for voice settings"
          },
          "language": {
            "type": "string",
            "description": "Language setting"
          }
        },
        "required": [
          "inputMessage"
        ]
      },
      "TextToSpeechTimestampedResponseDto": {
        "type": "object",
        "properties": {
          "audio": {
            "description": "Array of base64 encoded audio chunks",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "words": {
            "description": "Array of words spoken",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "wtimes": {
            "description": "Array of start times for words",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "wdurations": {
            "description": "Array of durations for words",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "audio",
          "words",
          "wtimes",
          "wdurations"
        ]
      },
      "FullReplyPlainResponseDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Text reply from assistant"
          },
          "speech": {
            "description": "Generated speech audio response",
            "allOf": [
              {
                "$ref": "#/components/schemas/TextToSpeechResponseDto"
              }
            ]
          }
        },
        "required": [
          "text",
          "speech"
        ]
      },
      "FullReplyTimestampedResponseDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string",
            "description": "Text reply from assistant"
          },
          "speech": {
            "description": "Generated timestamped speech audio response",
            "allOf": [
              {
                "$ref": "#/components/schemas/TextToSpeechTimestampedResponseDto"
              }
            ]
          }
        },
        "required": [
          "text",
          "speech"
        ]
      },
      "RealtimeVoiceDto": {
        "type": "object",
        "properties": {
          "sdpOffer": {
            "type": "string",
            "description": "SDP offer for WebRTC"
          },
          "personality": {
            "type": "object",
            "description": "Personality configuration"
          },
          "conversationRole": {
            "type": "object",
            "description": "Conversation role"
          },
          "language": {
            "type": "string",
            "description": "Language setting"
          },
          "scenario": {
            "type": "object",
            "description": "Scenario configuration"
          },
          "userProfile": {
            "type": "object",
            "description": "User profile information"
          }
        },
        "required": [
          "sdpOffer"
        ]
      },
      "WebRtcAnswerResponseDto": {
        "type": "object",
        "properties": {
          "sdp": {
            "type": "string",
            "description": "WebRTC SDP answer"
          }
        },
        "required": [
          "sdp"
        ]
      },
      "RealtimeTranscriptionDto": {
        "type": "object",
        "properties": {
          "language": {
            "type": "string",
            "description": "Language setting"
          }
        }
      },
      "TurnDetectionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "threshold": {
            "type": "number"
          },
          "prefix_padding_ms": {
            "type": "number"
          },
          "silence_duration_ms": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "threshold",
          "prefix_padding_ms",
          "silence_duration_ms"
        ]
      },
      "InputAudioTranscriptionDto": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "prompt": {
            "type": "string"
          }
        },
        "required": [
          "model",
          "language",
          "prompt"
        ]
      },
      "ClientSecretDto": {
        "type": "object",
        "properties": {
          "expires_at": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "expires_at",
          "value"
        ]
      },
      "TranscriptionSessionCreateResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "modalities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "turn_detection": {
            "$ref": "#/components/schemas/TurnDetectionDto"
          },
          "input_audio_format": {
            "type": "string"
          },
          "input_audio_transcription": {
            "$ref": "#/components/schemas/InputAudioTranscriptionDto"
          },
          "client_secret": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ClientSecretDto"
              }
            ]
          },
          "expires_at": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "object",
          "modalities",
          "turn_detection",
          "input_audio_format",
          "input_audio_transcription"
        ]
      },
      "AiProviderStatusDto": {
        "type": "object",
        "properties": {
          "apiKey": {
            "type": "string",
            "description": "API Key enumerator"
          },
          "isAvailable": {
            "type": "boolean",
            "description": "Whether the API is available / configured"
          }
        },
        "required": [
          "apiKey",
          "isAvailable"
        ]
      }
    }
  }
}